Enable Travis + Sauce Labs integration
[tests] Update test script to support Travis + Sauce Labs
-----------------------------------------------------------

Also, fixed an issue where the wrong keyboard shortcuts were being sent
in the tests. We check the platform in order to decide on key sequences;
in case we are running remote, we want the platform that the client is
running on, not that of the build server.

SD-446 #resolve

[travis] Update Travis config to support Sauce Labs integration
---------------------------------------------------------------

- Our test script is not compatible with Node.JS 0.10; I changed the
  version to 0.12.4.

- Downloads binaries for MongoDB 3.0.6 (the Travis environment uses by
  default a very old version of Mongo that SlamEngine does not support).

- Enables Sauce Connect add-on (for setting up a tunnel)

-----------------------------------------------------------------

Also, moved Test.Env into the correct directory.